requirements.txt
hata
chardet>=2.0
PyNaCl>=1.3.0
python-dateutil>=2.0
Flask
sqlalchemy
python-dateutil
beautifulsoup4
Flask-WTF
WTForms
pillow
psutil
lxml
youtube_dl
